feat: implement Bar component
