+++ apps/dashboard/app/components/bars/ServerBar.tsx
@@ -1,11 +1,6 @@
 import React from "react";
-import { cn } from "~/lib/utils";
-
-interface BarProps {
-  value: number;
-  max: number;
-  className?: string;
-}
+import { getPercentage } from "~/lib/utils";
+import Bar from "./Bar";
 
 interface ResourceBarProps {
   title: string;
@@ -14,37 +9,20 @@ interface ResourceBarProps {
   max: number;
 }
 
-const Bar: React.FC<BarProps> = ({ value, max, className }) => {
-  const percentage = Math.min((value / max) * 100, 100);
-
-  const getBarColor = () => {
-    if (percentage > 90) return "bg-rose-500";
-    if (percentage > 75) return "bg-amber-500";
-    return "bg-emerald-500";
-  };
-
-  return (
-    <div
-      className={cn(
-        "h-1.5 w-full overflow-hidden rounded-full",
-        "bg-slate-200 dark:bg-slate-800",
-        className
-      )}
-    >
-      <div
-        className={cn("h-full rounded-full", getBarColor())}
-        style={{ width: `${percentage}%` }}
-      />
-    </div>
-  );
-};
-
 const ServerBar: React.FC<ResourceBarProps> = ({
   title,
   uiValue,
   value,
   max,
 }) => {
+  const percentage = getPercentage(value, max);
+
+  const getBarColor = (percentage: number) => {
+    if (percentage > 90) return "bg-rose-500";
+    if (percentage > 75) return "bg-amber-500";
+    return "bg-emerald-500";
+  };
+
   return (
     <div className="space-y-1">
       <div className="flex items-center justify-between text-xs">
@@ -53,7 +31,11 @@ const ServerBar: React.FC<ResourceBarProps> = ({
           {uiValue}
         </span>
       </div>
-      <Bar value={value} max={max} />
+      <Bar
+        percentage={percentage}
+        barColor={getBarColor(percentage)}
+        size="md"
+      />
     </div>
   );
 };
